Added Flight Data
4 routes a day
Each primary key is a concantenation of the data field and the route field. Unique based on day, and by route, which hold the time data.

Pilots are routed in order and don't need any manipulation if the flight routes are kept in order 1 then 2 then 3 then 4. It all works out super elegantly.